requirements.txt
fastapi
uvicorn[standard]
pyinstaller
web3
pillow
pandas
